[closure] Subst variables in ClosureSubstSpecializedMethod
Summary: In `ClosureSubstSpecializedMethod`, it duplicates a procedure with specialized closure parameters. Since it introduces a new procedure name, its local variables in the procedure body must be replaced to use the new procedure name. (Note that local variable type includes procedure name.) However, in the previous implementation, it missed the translations in some cases: compound expressions and metadata.

Reviewed By: ezgicicek

Differential Revision: D26075490

fbshipit-source-id: 2a5a30cd8