[frontend] Add tests of using single field struct
Summary:
When a single field struct is initialized with "type x{v}" form, the translated result is not straightforward.  For example,

```
struct t {
  int val_;
};

void foo(t x) {
  t y{x};
}
```
calls the copy constructor with `x`.  This is good.

```
void foo(int n) {
  t y{n};
}
```
assigns the integer `n` to `y.val_`.  This is good.

```
t get_v();

void foo() {
  t y{get_v()};
}
```

assigns return value of `get_v` to `y.val_`, rather than calling the copy constructor.  This is not
good, but doesn't matter for actual running; `&y.val_` is the same to `&y` and `t` value is the same
to `int` value.

Reviewed By: jvillard

Differential Revision: D26146578

fbshipit-source-id: 8a81bb1db