[inferbo] Handle when const global assigned to local variable
Summary:
Before this diff:

```
// Summary of const global
// { global -> v }

n$0 =* global
// n$0 -> {global}
x *= n$0
// x -> {global}
```

However, this is incorrect because we expect `x` have `v` instead of the abstract location of `global`.

To fix the issue, this diff lookups the initializer summary when `global` is evaluated as RHS of load statement.

After this diff:

```
// Summary of const global
// { global -> v }

n$0 =* global
// n$0 -> v
x *= n$0
// x -> v
```

Reviewed By: ezgicicek

Differential Revision: D26369645

fbshipit-source-id: 98b1ed085