[racerd] races in nullsafe classes
Summary:
Races in Nullsafe classes can undermine NPE safety despite the class passing the type checks.

This diff adds to the report text of THREAD_SAFETY_VIOLATION and GUARDEDBY_VIOLATION the following trailer:

> Data races in `Nullsafe` classes may still cause NPEs.

This only happens if the race is directly on a non-primitively-typed member field of the class.

It also uses distinct bug types (adds the suffix _NULLSAFE to the bug types above) for easier accounting.

Reviewed By: ezgicicek

Differential Revision: D26403274

fbshipit-source-id: 3cd6ca082