[website] fix links to static resources
Summary:
Links to pages in static resources have been broken for a while it
seems: Docusaurus generates a Page not Found error for them, then
refreshing the URL loads the page properly... This is because of wrong
routing from the Docusaurus Single-Page App.

See https://github.com/facebook/docusaurus/issues/3309

A suggested workaround is to prepend `pathname://` in front of each link
not managed by docusaurus. This makes the page open in a new tab, which
is actually quite ok for our use cases: OCaml API docs and man pages.

Reviewed By: skcho

Differential Revision: D26544273

fbshipit-source-id: 2ed86ca4f