[ConfigImpact] Only report on functions that occur on a given json config data
Summary:
Currently, we report on all functions that are not config checked. However, the aim of the analysis is to only report on these for specific functions. Moreover, this has performance implications in practice.

This diff instead reports on functions that occur on a json file that is passed by the command line option `config-data-file`.

Reviewed By: skcho

Differential Revision: D26666336

fbshipit-source-id: 290cd3ada