[pulse][nullptr][objc] Add nil summaries for ObjC instance methods
Summary: When a method is called in ObjC on nil, there is no NPE, the method is actually not called and the return value is 0/false/nil. (There is an exception in the case where the return type is non-POD. In that case it's UB. This will be addressed later). To implement this behaviour we add additional summary to ObjC instance methods {self = 0} {return = 0}. We also want to make sure that inferred summary will not be used in we call a method on nil, hence, we add a path condition {self > 0} to get a contradiction when needed.

Reviewed By: jvillard

Differential Revision: D26664187

fbshipit-source-id: cdac2a5bb