[starvation] remove severity
Summary:
A "severity" level was used for keeping the highest severity issue when deduplicating across many issues on the same location. This was used for only one of the issue types reported by the analysis.

It turned out this isn't very useful and it complicates significantly the reporting code.

This diff removes the type and uses trace depth to sort all the issue types in the same way all other issue types than `starvation` used.

Next diff will remove now unnecessary stuff.

Reviewed By: skcho

Differential Revision: D26725569

fbshipit-source-id: f9287dcd1