[pulse][objcpp][nullptr] Link self with return to preserve the trace
Summary: We were missing a part of the trace if it was going through a nil summary as the invalidation was set in the nil summary. Instead of creating a fresh value for return in the nil summary {self=0}{return=0}, we return self {self=0}{return=self}. This way we keep all the information about invalidation in the trace.

Reviewed By: jvillard

Differential Revision: D26871098

fbshipit-source-id: 6eb175e68