[racerd] fix analysis of class initializers
Summary:
RacerD needs to analyse the class initialiser in order to establish field properties in its post, such as that certain static fields are synchronized containers.

There was a bug where class initializers were not analysed at all, from the time where there was no analysis of field properties in the post.

We still don't want to report on the class initialiser since it cannot possibly race with itself (JVM guarantees that) and it cannot race with any of the other methods in its class (because it must finish before any other method can be called).

Reviewed By: da319

Differential Revision: D26887151

fbshipit-source-id: 570aff370