[objc] Suppress block counter when getting method name
Summary:
We use `procedure_name` which is coming from `Procname.get_method` in explanation of cost issues. For blocks, procedure name includes a prefix `objc_block` and a suffix with `_x` where x is the block counter. However, displaying this name to the user is not pretty. Especially when we have nested blocks, procedure name looks like `objc_blockobjc_blockdirectUIMessageFromContentAndMetadata_10_23`.

This diff drops the block index suffix and replaces `objc_block` with a prettier version `^`(signifying block).

so instead, in the cost report, we will have `^^blockdirectUIMessageFromContentAndMetadata`.

Reviewed By: skcho

Differential Revision: D26945333

fbshipit-source-id: 9d135423c