[liveness] Handles live variables in catch block
Summary:
This diff handles live variables in catch blocks.  To do that, this diff adds another metadata,
`CatchEntry`.

Domain change: The domain is changed to

```
(normal:variables) x (exn:try_id->variables)
```

`exn` is a map from try-catch-id to a set of live variables that are live at the corresponding entry
of catch blocks.

Semantics change: It is a backward analysis.

* on `CatchEntry`: It updates `exn` with `try_id` and current `normal`.
* on `Call`: As of now, we assume all function calls can raise an exception. Therefore, it copies
  all live variables in `exn` to `normal`.
* on `TryEntry`: It removes corresponding `try_id` from `exn`.

Reviewed By: jvillard

Differential Revision: D26952755

fbshipit-source-id: 1da854a89