[pulse][2/2] generate latent issues when null is allocated
Summary:
See updated tests and code comments: this changes many arithmetic
operations to detect when a contradiction "p|->- * p=0" is about to be
detected, and generate a latent issue instead. It's hacky but it does
what we want. Many APIs change because of this so there's some code
churn but the overall end result is not much worse thanks to monadic
operators.

Reviewed By: skcho

Differential Revision: D26918553

fbshipit-source-id: da2abc652