[website] delete outdated "Advanced features" section
Summary:
Almost all of the information on this page is wrong :)

I moved some documentation on debug mode to the previous commit.

Debug information is present in CONTRIBUTING.md which is a better source
of truth. Anecdotally it's been kept more up to date than this web page
so far!

Reviewed By: skcho

Differential Revision: D27268171

fbshipit-source-id: 25314a823