[sledge][cxxabi] Vendor LLVM cxxabi project
Summary:
Add a vendored copy of the LLVM cxxabi project.

The sledge models compile llvm's libcxxabi to bitcode, so that the
analyzer knows the real definitions of the e.g. exception handling
primitives. Before this diff, the sledge build relied on having a
clone of a fork of llvm in the sledge working tree. This diff is part
of enabling using the upstream llvm 11 library instead of the fork.

Source: https://github.com/llvm/llvm-project/tree/main/libcxxabi

Reviewed By: jvillard

Differential Revision: D27188298

fbshipit-source-id: dc76b0714